transport/tcp: Improve substream protocol negotiation
Combine substream opening and negotiation events under the same future
to prevent signaling and setup overhead. The implementation is not
perfect because right now each substream is negotiated individually
which causes the Yamux connection to block until the substream has
completed negotiation. The solution is to store the negotiation futures
into a separate store which are polled alongside the `yamux::Connection`
object as not to disturb either process from progressing independently.

While the current implementation has its limitations, it is
architecturally a step in the right direction as there should be as
much parallelism in the code as possible while simultaneously keeping
the setup overhead as minimal as possible.