transport/webrtc: Sleep a little before handling the opened channel
There is a bug somewhere either in `litep2p` or `smoldot` where the first
part of the handshake fails when it's send too fast. This causes `smoldot`
to crash because the code attempts to handle a stream message before the
connection has been registered properly.

This needs further investigation to find out the root issue but sleeping
a bit before sending the initial handshake gives the browser enough time
to create the necessary context to handle the message.