transport/websocket: Introduce `BufferedStream`
`multistream-select`, `noise` and `yamux` expect the substream
to implement `futuresAsyncRead + futures::AsyncWrite` which the
WebSocket stream doesn't implement by default because it accepts data
in different types (binary, text, ping/pong).

Introduce a new buffered stream type that encapsulates the underlying
WebSocket connection and implements the necessary traits for it.