protocol/libp2p/kademlia: Introduce `QueryEngine`
`QueryEngine` is responsible for tracking the state of outbound queries,
schedule new outbound queries, determining when a query has failed or
succeeded. It works in tandem with `Kademlia` which handles the I/O
side of things and reports all inbound responses to `QueryEngine`.

The code is right now quite bad but can be improved later on