litep2p: Improve keep alive timeout handling
Introduce a new object called `Permit` which protocol must acquire when
they wish to open substreams or when an inbound substream is opened.

`Permit` allows the protocol to keep the connection open while the
substream is still active and when the substream is dropped, so is the
permit which makes it possible to kill the connection using keep-alive
timeout. If all permits are dropped, the connection notices this and
closes itself.

`Permit` can only be acquired if there is at least one other `Permit`
active and if all permits have been dropped, protocol/transport is
unable to acquire a new permit, forcing the connection to be closed.

When a connection is established, the first `Permit` (through
`ConnectionHandle`) is given to `ProtocolSet` and that `Permit` must be
manually dropped by calling `TransportService::disconnect()`.

One improvement left to be made is to allow the permit to be dropped
automatically after the initial keep-alive timeout is over and after
that timeout is over, the connection is kept alive only by having
substerams open over the connection.