protocol/request-response: Close substream on response error
Sending too large response will cause `send()` to fail but
the substream is not closed. Depending on whether the connection
is ready to be closed due to all connection permits having been dropped,
not closing the substream would result in different kinds of errors.

To have consistency in returned errors, always close the substream if
sending a response fails.