protocol: Add support for fallback protocols
If the protocol has changed its name and is now using a new name but
wishes to still accept substreams using the old name, that old name can
be provided as a fallback protocol and provided to `multistream-select`
when the protocol is negotiated.

If the substream was negotiated using a fallback name, the `fallback`
field in `TransportEvent::SubstreamOpened` is set to that fallback
protocol name.