protocol/request-response: Introduce `DialOptions`
User can now decide when sending a request to remote peer what should
happen if the peer is curently not connected. `DialOptions::Dial` will
dial the peer and if the connection is established, send the request to
peer.

If `DialOptions::Reject` is given to `send_request()` and the peer is
not connected, the request will fail and
`RequestResponseError::NotConnected` is reported to the user.