protocol/notification: Delay starting the connection handler
Start connection handler only after the open substream has been reported
to the user. Polkadot SDK has a bug where `ProtocolController` accepts
an inbound connection but this connection is later rejected by
`SyncingEngine` and the peer gets disconnected. Technically this is
a protocol violation which means that `litep2p` first gets
an acknowledgedment for the substream, only for the substream to be
closed immediately after.

If the connection handler is started before the open substream is
reported to user and the handler notices that the connection is closed,
the order of the events might get reserved, meaning that the user first
gets an event saying the notification stream has been closed, followed
by an event saying that the notification stream has been opened.