transport: Refactor `TransportManager`
Instead of storing peer addresses in a `HashSet` and always retrying
the first address, wrap `Multiaddr` in an `AddressRecord` which contains
a score for the address and store the address records in a `BinaryHeap`.

When connection is established, the address score is increased and when
dial fails, the score is decreased. If the address score falls below
a certain threshold, the address is considered dead and if the peer
doens't have any other addresses, the peer is considered undialable
and it can also be removed from any higher-level protocol's storage
that deals with selecting peers for outbound connections.