protocol/request-response: Limit the number of inbound requests
Limit for the maximum inbound requests can be set with
`ConfigBuilder::with_max_concurrent_inbound_requests()`. This will
limit the number of inbound requests (both reading the request and
waiting for a response from user) the request-response protocol may
have active any given moment in time.

Excess inbound requests are dropped by closing the substream which
indicates to the remote peer that the request was rejected.