protocol: Remove `debug_assert!()`
Nothing prevents remote from opening more than two connections and even
though the extra connections are ignored by the local node and will
get closed due to a keep-alive timeout expiring, event about the
connection closing is still reported which is not an invalid state
transition.

Whether all connections to node should be closed because it has opened
multiple connections even though it's invalid logic is something that
needs to be researched further.