protocol/notification: Report closed substream from `Connection`
Both `NotificationProtocol` and `Connection` reporting events caused
a race condition where `NotificationProtocol` might have detected that
the connection was closed but `NotificationHandle` hadn't consumed all
notifications and as the events from `Connection`/`NotificationProtocol`
were interleaved, user calling `NotificationHandle::next()` might first
read an event saying that connection was closed and then consume rest
of the notifications from the peer.

This would cause issues, e.g., in `SyncingEngine` where it would print
a warning saying that a block announcement was received from a rejected
peer (which was just a disconnected peer).