protocol/libp2p/identify: Remove `debug_assert!()`
Depending on how the tasks get polled, what can happen is that
connection is opened and it gets registered to `Identify` but the
protocol doesn't get polled until for enough time that the connection
is closed. This can cause the outbound substream to not be opened
successfully and the observed address is not inserted for the peer.

If `Identify` also received an event for an inbound substream for the
(now disconnected) peer, the observed address wouldn't be available,
causing the code to crash to the `debug_assert!()`