protocol/request-response: Track active requests correctly
Previously the code would insert pending outbound request to `active`
but forgot to remove them if substreams would failed to open.

This caused `RequestResponseProtocol` to emit `Rejected` for the same
request twice: once after the substream failed to open and again after
the peer disconnected.

Insert the pending outbound request to `active` only if the substream
is opened successfully