transpor/tcp: Introduce `TcpListener`
`TcpListener` allows the TCP transport to listen to zero or more
addresses simultaneously.

This enables starting a transport with no listen addresses in which
case it can only be used to dial other peers. It also allows users to
listen to multiple TCP addresses simultaneously.

Commit also prepares WebSocket/QUIC/WebRTC for this same refactoring