protocol/notification: Add timeout for opening outbound substreams
There is a bug in Substrate whereby `Notifications` may get confused
about the connection state since it cannot detect that it's closed
without sending a notification to remote peer. This means that if remote
has closed the connection, `Notifications` may not detect this and if
remote decides to open another substream, it gets automatically accepted.

However, because `Notifications` erroneously thinks the connection was
open, it doesn't open an outbound substream towards the remote peer and
thus the outbound connection attempt has to be bounded in time to allow
`NotificationProtocol` to notify whoever initiated the connection that
it has failed.