transport/tcp: Return `ConnectionEstablished` from `TcpTransport`
Refactor `TcpConnection::negotiate_connection()` to return
`NegotiatedConnection` which holds the active connection state but
doesn't start the connection event loop.

After the connection has been negotiated, return
`TransportEvent::ConnectionEstablished` to `TransportManager` which
validates the connection and if it's accepted, event loop for the
negotiated connection is started. Protocols are notified of the
connection only after the event loop for the connection is started.

This commit also introduced two new functions to `Transport`:
  - `Transport::accept()`
  - `Transport::reject()`

which are called by `TransportManager` after the connection has been
validated to either start the event loop (`accept()`) or drop the
connection (`reject()`).

Currently only `TcpTransport` supports this functionality