transport: Allow dialing peer over multiple addresess in parallel
Introduce three new functions to `Transport`:
  - `Transport::open()`
  - `Transport::negotiate()`
  - `Transport::cancel()`

The first function is used to initiate a connection with remote peer
over one or more addresses. This will result in the transport only
dialing the peer and if connection is established over any of the
addresses, the event is reported to `TransportManager`.
If the connection is accepted by `TransportManager`, it calls
`Transport::negotiate()` which actually negotiates the stream paramters
(noise/multistream-select) and starts the connection.

This functionality allows dialing a peer over multiple addresses and
transports at the same time. If for one peer there are three known TCP
addresses, two WebSocket addresses and 1 QUIC address,
`TransportManager` will attempt to connect to the peer using all those
addresses in parallel.