protocol/notification: Make dialing configurable
The default behavior of the notification protocol is that if an outbound
substream request is made and the peer is not connected, it will be
dialed. This is not appropriate behavior for some protocols that depend
on other protocols, such as syncing and GRANDPA in Polkadot SDK.

Add ability to disable the default behavior and error out early if the
peer is not connected.