transport: Make dialing atomic
Since dialing using a `PeerId` was split into two functions, `dial()`
and `dial_inner()` there was the possibility that the first protocol
dialing a peer would drain all addresses from peer's `AddressStore`
while another protocol was waiting on the lock. Once the lock was
released and the first protocol would proceed to `dial_inner()` the
second protocol would call `dial()` and if it acquired the lock before
the first protocol had acquired the lock again in `dial_inner()`,
the second dial attempt would fail with `NoAddressesAvailable` even
though a dial attempt was already in progress. This would then be
reported, falsely, as a notification protocol open failure.

Combine `dial()` and `dial_inner()` to make sure that peer state is
valid even if two protocols attempt to dial the same peer at the same
time.