protocol/request-response: Remove substream close detection
Polkadot SDK closes the writer half of the substream which was detected
by litep2p incorrectly as the whole substream getting closed. Currently
there is some undiagnosed problem with yamux as it doesn't report an
error when trying to send to a closed substream.

This will result in some false-positive feedbacks but it will be fixed
in the near future.