protocol/notification: Report outbound open failures to user
If an inbound substream was accepted but the corresponding outbound
substream failed to open, report the failure to user so they can track
the peer states correctly.

Originally the code was correct but the code tracking peer states on
Polkadot SDK side has changed since the initial implementation and it
depends on receiving this event.