protocol: Introduce `ProtocolCommand::ForceClose`
Sometimes it may be necessary to forcibly close the connection, e.g.,
due to a protocol in violation in one protocol that deserved the whole
connection to be closed. Protocol can cal
`TransportService::force_close()` to forcibly terminate the connection
even if the keep-alive timeout is still pending or if other protocols
have substreams open over the connection.

Force closing a connection should be used sparingly since litep2p should
not be the one that decides which connections are open and which are
closed, unless the connection used, which again is a signal from the
user protocol that it is not interested in the connection.

`ProtocolCommand::ForceClosed` is also used in the notification protocol
to work around a known issue in Polkadot SDK's `Notifications`
implementation whereby it can have a pair of uncorrelated substreams
due to a bug in the implementation.