protocol/request-response: Add support for fallback requests
Add support for binary-incompatible fallbacks. If user sends a request
using `RequestResponseHandle::send_request_with_fallback()` and the
negotiated protocol is the one that is considered the fallback,
`RequestResponseProtocol` will send that request instead of the default
request and if response is received from remote for the fallback request,
the fallback protocol is reported in `ResponseReceived` event.