protocol/notification: Handle missing substreams gracefully
If inbound substream succeeded to negotiation, it was pushed to `ready`
and if the outbound substream then failed to negotiate, an error was
returned immediately for the outbound substream and
`NotificationProtocol` would remove both substreams from
`HandshakeService`. The inbound substream was still left in `ready`
and next time `HandshakeService` was called, it would try pop the event
from `ready` and an invariant in the code assumed that since an event
was produced by an substream (item in `ready`), that a substream for
that peer would also exist which is not the case considering how
`NotificationProtocol` works.

Handle missing substreams gracefully by looping over `ready` until it's
empty or an existing substream is found.