transport/webrtc: Refactor WebRTC connection opening
Introduce `OpeningWebRtcConnection` which only deals with handling the
initial Noise handsdhake which starts every WebRTC connection.

When `str0m` receives an inbound connection from a new peer,
an `OpeningWebRtcConnection` object is created which tracks the Noise
handshake state of the connection. This object doesn't have its own
asynchronous event loop and is polled by the `WebRtcTransport` object.

Once the remote `PeerId` has been received in the second Noise message,
`TransportManager` will be notified of the new connection which it
validates and either accepts or reject. If the connection is rejected,
all connection state is discarded.

If the connection is accepted, `OpeningWebRtcConnection` sends the third
and final Noise handshake message, fully opening the connection.
After that, the object is destroyed and a `WebRtcConnection` object
is crated using the `Rtc` object that was held by
`OpeningWebRtcConnection`. This connection object has its own
asynchronous event loop like connection loops of other transports.

Currently the `WebRtcConnection` is just a dummy connection which polls
the connection and negotiates protocols for inbound substreams but does
nothing else. Proper connection support will be implemented in future
commits.