protocol/libp2p/kademlia: Keep track of all parallel queries
Previously the code would construct `PeerContext` in such a way that
if there were multiple pending actions for a peer that was being dialed,
only the most recent action was stored and others were silently
discarded. Code didn't consider the fact that there might already be
pending actions for the peer and essentially overwrote previous actions.
Once the result of the dial was received from the transport
(succcess/fail), only the most recent action was considered and previous
actions were discarded, leaving the original query in a state where it
can't make progress.

When handling pending actions, make sure to process all of them and
correctly add new actions in case there already are pending actions