litep2p: Add yamux source codes to litep2p
yamux `0.11.1` and `0.12.1` contains fixes for how its GC works and
while upgrading to `0.11.1` would be trivial, the fix was not enough not
enough to reduce memory leakage. `0.12.1`, on the other hand, contians
significant API breakage and likely requires a rewrite of TCP/WebSocket
connection handlers as `ControlledConnection` and `Control` were removed

In the intersest of saving time, switch to `0.11.1`, backport the GC fix
from `0.12.1` to `0.11.1`, and move source code of yamux to litep2p.

This is by no means an ideal solution but considering that fixing the
memory leaks are of highest importance and rewriting the connection
handlers likely requires significant rewrite and extensive testing,
it's easiest to just use yamux directly, like was done with
multistream-select.