Revert "Bring `rustfmt.toml` in sync with polkadot-sdk (#71)" (#74)
This reverts commit 0ca33beaf15d59e48f5edd88ed5791ab8f38dd91.

It turned out to be a bad idea to apply major code style changes while
having unmerged PRs. Specifically, in
https://github.com/paritytech/litep2p/pull/51 I hit the all file long
merge conflicts where it's impossible to manually track important
changes and merge them. Applying the new code style to the PR itself
also didn't help — git failed to correctly handle the changed parts and
this led to even more merge conflicts.

So, the plan is to revert the code style changes, apply the pending PRs,
and reintroduce the code style changes.