transport/manager: Add connection limits for inbound and outbound established connections (#185)
This PR adds connection limits for: 
- maximum number of inbound established (negotiated) connections
- maximum number of outbound established (negotiated) connections
- any dial attempts are immediately rejected if we reach capacity for
the maximum number of outbound established connections

The public API exposes for the connection limits a builder, both on the
Litep2pConfig, as well for the limits:

```rust
let litep2p_config = Config::default()
 .with_connection_limits(ConnectionLimitsConfig::default()
      .max_incoming_connections(Some(3))
      .max_outgoing_connections(Some(2)));
```

The connection limits increments and decrements connections identified
by their `ConnectionId(usize)`.
Regarding memory consumption, we use 2 hash-maps that require
`sizeof(usize) * (num_of_inbound + num_of_outbound)` for elements.
Where,`num_of_inbound` may be capped to `max_incoming_connections` and
`num_of_outbound` to `max_outgoing_connections`.

We do not need to introduce a limit on the total number of connections
established with a single peer. Litep2p already assumes we can have a
maximum of 2 established connections for each peer.

We could also limit the number of inbound non negotiated connections,
similar to how we handle outbound connections via dial methods. However,
each protocol eagerly accepts and negotiates all inbound connections.
This is not optimal, because we can save resources (CPU + mem) by
rejecting non negotiated connections when we are at inbound capacity.

To achieve this, a refactoring needs to move the accepting of inbound
connections back into the ownership of the TransportManger, instead of
the Transport itself. This PR introduces some changes, to also make it
easier for review will look into it as a follow-up:
- https://github.com/paritytech/litep2p/issues/186

Part of:
- https://github.com/paritytech/litep2p/issues/17

---------

Signed-off-by: Alexandru Vasile <alexandru.vasile@parity.io>