query-executor: Save the task waker on empty futures (#219)
This PR ensures that we save the task `waker` when the executor does not
have any pending futures (messages queued to send / receive from the
individual substreams).

When the `self.futures` are empty we returned `Poll::Pending`.
However, the given `QueryExecutor::poll_next` would never wake up again.
To mitigate this behavior, save the current waker and notify it upon
modifications to `self.futures`.

This was working before because we relied on other tasks to make
progress via the `loop { tokio::select! }`. However this ensures that
the query executor is efficiently waking up and can work in isolation.

cc @paritytech/networking

---------

Signed-off-by: Alexandru Vasile <alexandru.vasile@parity.io>