kad: Improve robustness of addresses from the routing table (#369)
This PR improves the robustness of the kademlia addresses stored in the
routing table.

The implementation uses an eventually-consistent address store, based on
the information provided by the transport manager. This has the benefit
of not sharing the transport manager address store with the routing
table, therefore not paying for the performance hit of locking the
addresses (globally) on kademlia queries.

- transport manager exposes all dialed addresses properly to the
installed protocols
- the kademlia peer contains an address store instead of a plain vector
(bounded to 32 entries -- please note that the transport manager holds
up to 64 entries)
- similar to the transport manager, the addresses are sorted by score
depending on success / failure dials
  - when an address score is negative it is subject to removal
- on successful dial, the score is increased for the address of said
peer
- on dial failures, the score of the failed addresses is decreased
- known good addresses are always provided first
- all kademlia queries rely on this information and the addresses are
provided sorted by their score

This resolved the following issue, with the mention that undialed or
dialed but failed addresses are still provided if the address store has
enough capacity. This ensures that addresses that fail due to temporary
network errors, or errors specific to our node (ie, behind firewall --
other peers may reach the address) are provided:
- https://github.com/paritytech/litep2p/issues/209

cc @paritytech/networking

---------

Signed-off-by: Alexandru Vasile <alexandru.vasile@parity.io>
Co-authored-by: Dmitry Markin <dmitry@markin.tech>