ci: Update stable dependencies (#375)
This PR is part of our ongoing efforts to update and clean up the
dependencies for the stable feature set. Below are the key updates:

### The following crate dependencies are updated:
- ring to 0.17.14
- serde_json to 1.0.140
- bytes to 1.10.1
- prost to 0.13.5
- tokio-util to 0.7.14
- indexmap to 2.9.0
- thiserror to 2.0.12
- socket2 dep to 0.5.9
- smallvec to 0.15.0
- async-trait dep to 0.1.88
- tokio to 1.44.2
- serde dep to 1.0.219
- pin-project dep to 1.1.10
- tokio-tungstenite dep to 0.26.2
- cid dep to 0.11.1

### rcgen is dependant on Quic

The `rcgen` crate was wrongly left in the `Cargo.toml` as a mandatory
dependency. Considering this is part of the experimental `quic` feature
flag, it is now optional.

---------

Signed-off-by: Alexandru Vasile <alexandru.vasile@parity.io>