crypto/noise: Set timeout limits for the noise handshake (#373)
This PR sets timeout limits for the crypto/noise handshake negotiation.

By default, the timeout is bounded similar to protocol negotiation to
the substream open timeout.

To allow this change, the WebSocket internal API has been opened to
accept the open timeout. Before this change, we had no granular control
and used the connection timeout as a fallback.

cc @paritytech/networking

---------

Signed-off-by: Alexandru Vasile <alexandru.vasile@parity.io>