yamux/control: Ensure poll next inbound is not called after errors (#445)
We have recently fixed a spurious issue that caused the rust-yamux
component to panic.

Litep2p transitioned the internal state of the connection back to Idle.
This causes litep2p to poll the component again. Then rust-yamux would
use the same buffer offset that caused the panic in the first place.

Instead, this PR ensures that litep2p never polls the connection once
yamux encounters an error.
This is aligned with the behavior of transport layers that close the
connection on the first yamux error.

Part of: https://github.com/paritytech/polkadot-sdk/issues/9169

cc @paritytech/networking

---------

Signed-off-by: Alexandru Vasile <alexandru.vasile@parity.io>